backend/extract/requirements.txt
psycopg2-binary
boto3
Pillow
PyMuPDF
